bot/requirements.txt
discord.py
python-dotenv
jishaku
requests
pyjwt
cryptography
